Fixed Checks

+++ akrolyb/src/main/java/com/jaqxues/akrolyb/pack/PackMetadata.kt
@@ -12,5 +12,5 @@ interface PackMetadata {
     val packVersion: String
     val packVersionCode: Int
     val packImplClass: String
-    val minApkVersionCode: Long
+    val minApkVersionCode: Int
 }
